refactor(core): reorg routes
