+++ packages/core/src/routes/admin-user/search.ts
@@ -6,7 +6,7 @@ import koaGuard from '#src/middleware/koa-guard.js';
 import koaPagination from '#src/middleware/koa-pagination.js';
 import { parseSearchParamsForSearch } from '#src/utils/search.js';
 
-import type { AuthedRouter, RouterInitArgs } from './types.js';
+import type { AuthedRouter, RouterInitArgs } from '../types.js';
 
 export default function adminUserSearchRoutes<T extends AuthedRouter>(
   ...[router, { queries }]: RouterInitArgs<T>
