+++ packages/core/src/routes/admin-user/social.ts
@@ -12,7 +12,7 @@ import RequestError from '#src/errors/RequestError/index.js';
 import koaGuard from '#src/middleware/koa-guard.js';
 import assertThat from '#src/utils/assert-that.js';
 
-import type { AuthedRouter, RouterInitArgs } from './types.js';
+import type { AuthedRouter, RouterInitArgs } from '../types.js';
 
 export default function adminUserSocialRoutes<T extends AuthedRouter>(
   ...[router, tenant]: RouterInitArgs<T>
